+++ python/paddle/fluid/tests/unittests/collective/fleet/pipeline_mnist_multi_device.py
@@ -105,7 +105,7 @@ class TestDistMnist2x2(TestDistRunnerBase):
 
         # Evaluator
         with fluid.device_guard("gpu:1"):
-            batch_size_tensor = fluid.layers.create_tensor(dtype='int64')
+            batch_size_tensor = paddle.tensor.create_tensor(dtype='int64')
             batch_acc = paddle.static.accuracy(
                 input=predict, label=label, total=batch_size_tensor
             )
