+++ python/paddle/fluid/tests/unittests/dist_fleet_raw_program_optimizer_fuse_allreduce.py
@@ -83,7 +83,7 @@ class TestFleetMetaOptimizerFuseAllReducePrecision(TestDistRunnerBase):
         avg_cost = paddle.mean(x=cost)
 
         # Evaluator
-        batch_size_tensor = fluid.layers.create_tensor(dtype='int64')
+        batch_size_tensor = paddle.tensor.create_tensor(dtype='int64')
         batch_acc = paddle.static.accuracy(
             input=predict, label=label, total=batch_size_tensor
         )
