+++ python/paddle/fluid/tests/unittests/test_desc_clone.py
@@ -77,7 +77,7 @@ def get_model(batch_size):
     avg_cost = paddle.mean(x=cost)
 
     # Evaluator
-    batch_size_tensor = fluid.layers.create_tensor(dtype='int64')
+    batch_size_tensor = paddle.tensor.create_tensor(dtype='int64')
     batch_acc = paddle.static.accuracy(
         input=predict, label=label, total=batch_size_tensor
     )
