+++ python/paddle/tensor/__init__.py
@@ -19,6 +19,8 @@ from .attribute import shape  # noqa: F401
 from .attribute import real  # noqa: F401
 from .attribute import imag  # noqa: F401
 from .attribute import is_floating_point  # noqa: F401
+from .creation import create_parameter  # noqa: F401
+from .creation import create_tensor  # noqa: F401
 from .creation import to_tensor  # noqa: F401
 from .creation import diag  # noqa: F401
 from .creation import diagflat  # noqa: F401
@@ -289,6 +291,8 @@ from .einsum import einsum  # noqa: F401
 
 # this list used in math_op_patch.py for _binary_creator_
 tensor_method_func = [  # noqa
+    'create_parameter',
+    'create_tensor',
     'matmul',
     'dot',
     'cov',
